Move changeFormView and disPlayTrips functions to domManipulation file
